Adding in unit testing for graph building
