+++ cmd/graph/dir_domain_emitter.go
