+++ test/test-graphbuild-gdc.yaml
@@ -1,0 +1,16 @@
+
+
+graphMap: examples/gdc/graph-map.yaml
+
+inputDir: examples/gdc/out/
+
+outputs:
+ - out-graph/Aliquot.Aliquot.Vertex.json.gz
+ - out-graph/Project_Case.cases.Edge.json.gz
+ - out-graph/Aliquot_Sample.derived_from.Edge.json.gz
+ - out-graph/Sample_Aliquot.aliquots.Edge.json.gz
+ - out-graph/Case.Case.Vertex.json.gz
+ - out-graph/Sample_Case.derived_from.Edge.json.gz
+ - out-graph/Case_Project.member_of.Edge.json.gz
+ - out-graph/Sample.Sample.Vertex.json.gz
+ - out-graph/Case_Sample.samples.Edge.json.gz
